refactor: Request, Response, Domain 패키지 분리
