+++ app/src/main/java/nextstep/jwp/RequestHandler.java
@@ -6,8 +6,8 @@ import java.io.OutputStream;
 import java.net.Socket;
 import java.util.Objects;
 import nextstep.jwp.controller.Controller;
-import nextstep.jwp.model.Request;
-import nextstep.jwp.model.Response;
+import nextstep.jwp.model.reponse.Response;
+import nextstep.jwp.model.request.Request;
 import nextstep.jwp.utils.RequestAssembler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
