+++ app/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -4,10 +4,10 @@ import java.io.IOException;
 import java.util.Objects;
 import nextstep.jwp.model.FileType;
 import nextstep.jwp.model.PathType;
-import nextstep.jwp.model.Request;
-import nextstep.jwp.model.Response;
-import nextstep.jwp.model.Session;
-import nextstep.jwp.model.User;
+import nextstep.jwp.model.domain.User;
+import nextstep.jwp.model.reponse.Response;
+import nextstep.jwp.model.request.Request;
+import nextstep.jwp.model.request.Session;
 import nextstep.jwp.service.LoginService;
 
 public class LoginController extends AbstractController {
