+++ app/src/main/java/nextstep/jwp/db/InMemoryUserRepository.java
@@ -3,7 +3,7 @@ package nextstep.jwp.db;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
-import nextstep.jwp.model.User;
+import nextstep.jwp.model.domain.User;
 
 public class InMemoryUserRepository {
