+++ app/src/main/java/nextstep/jwp/exception/FileNotFoundException.java
@@ -2,7 +2,9 @@ package nextstep.jwp.exception;
 
 public class FileNotFoundException extends RuntimeException {
 
-    public FileNotFoundException(String message) {
-        super(message);
+    private static final String MESSAGE = "파일을 찾을 수 없습니다.";
+
+    public FileNotFoundException() {
+        super(MESSAGE);
     }
 }
