+++ app/src/main/java/nextstep/jwp/exception/LoginException.java
@@ -2,7 +2,9 @@ package nextstep.jwp.exception;
 
 public class LoginException extends RuntimeException {
 
-    public LoginException(String message) {
-        super(message);
+    private static final String MESSAGE = "잘못된 로그인입니다.";
+
+    public LoginException() {
+        super(MESSAGE);
     }
 }
