+++ app/src/main/java/nextstep/jwp/exception/RegisterException.java
@@ -2,7 +2,9 @@ package nextstep.jwp.exception;
 
 public class RegisterException extends RuntimeException {
 
-    public RegisterException(String message) {
-        super(message);
+    private static final String MESSAGE = "파일을 찾을 수 없습니다.";
+
+    public RegisterException() {
+        super(MESSAGE);
     }
 }
