+++ app/src/main/java/nextstep/jwp/model/request/Session.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.model.request;
 
 import java.util.HashMap;
 import java.util.Map;
