+++ app/src/main/java/nextstep/jwp/service/RegisterService.java
@@ -3,8 +3,8 @@ package nextstep.jwp.service;
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.RegisterException;
-import nextstep.jwp.model.Request;
-import nextstep.jwp.model.User;
+import nextstep.jwp.model.domain.User;
+import nextstep.jwp.model.request.Request;
 
 public class RegisterService {
 
@@ -22,14 +22,14 @@ public class RegisterService {
             User user = new User(account, password, email);
             InMemoryUserRepository.save(user);
         } catch (IndexOutOfBoundsException | NullPointerException exception) {
-            throw new RegisterException("잘못된 회원가입입니다.");
+            throw new RegisterException();
         }
     }
 
     private void validateDuplicate(String account, String email) {
         if (InMemoryUserRepository.existsByAccount(account)
                 || InMemoryUserRepository.existsByEmail(email)) {
-            throw new RegisterException("잘못된 회원가입입니다.");
+            throw new RegisterException();
         }
     }
 }
